refactor(enums): update ENUM structure
