+++ src/main/java/com/laporeon/expensetracker/enums/Role.java
@@ -1,5 +1,6 @@
-package com.laporeon.expensetracker.entities;
+package com.laporeon.expensetracker.enums;
 
 public enum Role {
-    ADMIN, USER
+    ADMIN,
+    USER
 }
